requirements.txt
# ==================================
# Scraper Dependencies
# ==================================
tweepy>=4.14.0
praw>=7.7.0
atproto>=0.0.50
apscheduler>=3.10

# ==================================
# Generator Dependencies (AI Models)
# ==================================
openai>=1.40.0
anthropic>=0.39.0
google-genai>=0.1.0
groq>=0.9.0

# ==================================
# Shared Dependencies
# ==================================
sqlalchemy>=2.0
python-dotenv>=1.0
pandas>=2.0
httpx>=0.27.0
tiktoken>=0.7.0
tenacity>=9.0.0
